# backend/swagger-docs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/comments": {
      "post": {
        "operationId": "CommentsController_create",
        "summary": "创建评论",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "评论创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          }
        },
        "tags": [
          "comments"
        ]
      },
      "get": {
        "operationId": "CommentsController_findAll",
        "summary": "获取评论列表",
        "parameters": [
          {
            "name": "product_id",
            "required": false,
            "in": "query",
            "description": "商品ID筛选",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Comment"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "comments"
        ]
      }
    },
    "/comments/{id}": {
      "get": {
        "operationId": "CommentsController_findOne",
        "summary": "根据ID获取评论详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "评论ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "404": {
            "description": "评论不存在"
          }
        },
        "tags": [
          "comments"
        ]
      },
      "patch": {
        "operationId": "CommentsController_update",
        "summary": "更新评论",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "评论ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "404": {
            "description": "评论不存在"
          }
        },
        "tags": [
          "comments"
        ]
      },
      "delete": {
        "operationId": "CommentsController_remove",
        "summary": "删除评论",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "评论ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "删除成功"
          },
          "404": {
            "description": "评论不存在"
          }
        },
        "tags": [
          "comments"
        ]
      }
    },
...
    "/admin/dashboard/sales-trend": {
      "get": {
        "operationId": "DashboardController_getSalesTrend",
        "summary": "获取销售趋势",
        "description": "获取指定周期内的销售趋势数据",
        "parameters": [
          {
            "name": "period",
            "required": false,
            "in": "query",
            "description": "统计周期：day(日), week(周), month(月)",
            "schema": {
              "enum": [
                "day",
                "week",
                "month"
              ],
              "type": "string"
            }
          },
          {
            "name": "days",
            "required": false,
            "in": "query",
            "description": "统计天数，默认30天",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功"
          },
          "401": {
            "description": "未授权"
          }
        },
        "tags": [
          "后台管理-数据大屏"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/dashboard/category-distribution": {
      "get": {
        "operationId": "DashboardController_getCategoryDistribution",
        "summary": "获取分类分布",
        "description": "获取商品分类分布统计数据",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取成功"
          },
          "401": {
            "description": "未授权"
          }
        },
        "tags": [
          "后台管理-数据大屏"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/dashboard/top-products": {
      "get": {
        "operationId": "DashboardController_getTopProducts",
        "summary": "获取热门商品",
        "description": "获取热门商品排行数据",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "限制数量，默认10",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功"
          },
          "401": {
            "description": "未授权"
          }
        },
        "tags": [
          "后台管理-数据大屏"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/dashboard/user-growth": {
      "get": {
        "operationId": "DashboardController_getUserGrowth",
        "summary": "获取用户增长",
        "description": "获取用户增长趋势数据",
        "parameters": [
          {
            "name": "days",
            "required": false,
            "in": "query",
            "description": "统计天数，默认30天",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功"
          },
          "401": {
            "description": "未授权"
          }
        },
        "tags": [
          "后台管理-数据大屏"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/dashboard/real-time-data": {
      "get": {
        "operationId": "DashboardController_getRealTimeData",
        "summary": "获取实时数据",
        "description": "获取系统实时运营数据",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取成功"
          },
          "401": {
            "description": "未授权"
          }
        },
        "tags": [
          "后台管理-数据大屏"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "二手交易系统 API",
    "description": "二手交易系统后端接口文档，包含前台用户接口和后台管理接口",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "number",
            "description": "商品ID",
            "example": 1
          },
          "user_id": {
            "type": "number",
            "description": "评论用户ID",
            "example": 1
          },
          "parent_id": {
            "type": "number",
            "description": "父评论ID（用于回复）",
            "example": 1
          },
          "content": {
            "type": "string",
            "description": "评论内容",
            "example": "这个商品看起来不错！"
          }
        },
        "required": [
          "product_id",
          "user_id",
          "content"
        ]
      },
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "评论ID"
          },
          "product_id": {
            "type": "number",
            "description": "商品ID"
          },
          "user_id": {
            "type": "number",
            "description": "评论用户ID"
          },
          "parent_id": {
            "type": "number",
            "description": "父评论ID（用于回复）"
          },
          "content": {
            "type": "string",
            "description": "评论内容"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间"
          }
        },
        "required": [
          "id",
          "product_id",
          "user_id",
          "content",
          "created_at",
          "updated_at"
        ]
      },
      "UpdateCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "评论内容"
          }
        }
      },
...
    }
  }
}